Separate waterfall steps and add utils
