+++ lib/stubs.js
@@ -1,4 +1,3 @@
-const buildWaterfallTracer = require('./waterfall_tracer');
 const noop = function() {};
 const returnEmptyLogger = function() { return emptyLogger; };
 const returnValue = function() { return 1; };
@@ -20,10 +19,6 @@ var emptyTracer = {
   inject: noop,
   extract: returnNull,
   Tags: {}
-};
-
-emptyTracer.helpers = {
-  waterfallTracer: buildWaterfallTracer(emptyTracer)
 };
 
 const emptyLogger = {
