+++ lib/tracer.js
@@ -8,7 +8,6 @@ const tracerFactory = require('./tracer_factory');
 const constants = require('./constants');
 const os = require('os');
 const tracerUtils = require('./tracer_utils');
-const buildWaterfallTracer = require('./waterfall_tracer');
 
 const protoRoot = new protobuf.Root();
 protoRoot.loadSync(path.join(__dirname, 'trace_context.proto'));
@@ -219,8 +218,7 @@ module.exports = function Tracer(agent, pkg, env, params) {
 
   obj.helpers = {
     wrapRequest: requestHelper(obj),
-    mapToTags: tracerUtils.mapToTags,
-    waterfallTracer: buildWaterfallTracer(obj, logger)
+    mapToTags: tracerUtils.mapToTags
   };
 
   return obj;
